+++ R/data.R
@@ -9,3 +9,11 @@
 #' }
 #' @source \url{https://glints.com/}
 "glintsjobid"
+
+#' Available fields for each dataset
+#'
+#' Comparison table describing available fields between dataset created.
+#'
+#' @format A data frame with \code{data.frame} format
+#'
+"fields_available"