+++ NEWS.md
@@ -1,3 +1,11 @@
+# unmplymnt 0.1.2
+
+* Added `kalibrr()` function.
+* Inactivated `indeed()` function.
+* Rearranged job vacancy result table's fields order.
+* Experimental `desc_glints()` to get job description.
+* Added `fields_available` dataset for fields comparison across job post providers.
+
 # unmplymnt 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.